Updated queries.js by importing get_user() and adding the logic to get the rubrics for the logged in user and any rubrics their owners have made! Updated AddCustomRubric.js by moving the CollapsableTableComponent into its own file and fixed logic for creating a rubric and displaying selected categories! Updated AdminAssCustomRubric.js by removing no longer needed import! Created CollapsableRubricCategoryTable.js and added Jerrys code!
