+++ FrontEndReact/src/View/Admin/Add/AddCustomRubric/AdminAddCustomRubric.js
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
 import "bootstrap/dist/css/bootstrap.css";
 import ErrorMessage from "../../../Error/ErrorMessage";
-import AdminAddCustomRubricView from "./AdminAddCustomRubricView";
 import { genericResourceGET, parseCategoriesToContained, parseCategoryIDToCategories } from '../../../../utility.js';
 import AddCustomRubric from "./AddCustomRubric";
